// packages/epi2melabs-wfpage/epi2melabs_wfpage-0.1.6.tar.gz/epi2melabs_wfpage-0.1.6/src/components/workflow/schema.ts
export type ParameterSection = {
  title: string;
  description: string;
  type?: string;
  fa_icon?: string;
  properties: { [key: string]: Parameter };
};

export type Parameter = {
  type: string;
  description?: string;
  help_text?: string;
  hidden?: boolean;
  default?: any;
  [key: string]: any;
};

export type WorkflowDefaults = {
  [key: string]: string;
};

export type WorkflowSchema = {
  [key: string]: any;
};

export type Workflow = {
  url: string;
  name: string;
  path: string;
  desc: string;
  schema: WorkflowSchema;
  defaults: WorkflowDefaults;
};